Refactor package structure to put acceptance package above stream
