+++ money-akka/src/test/scala/com/comcast/money/akka/acceptance/stream/TracedFlowSpec.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.comcast.money.akka.stream.acceptance
+package com.comcast.money.akka.acceptance.stream
 
 import akka.stream.Attributes
 import akka.stream.scaladsl.{ Keep, Sink, Source }
